fix: servingForをappSubtitleに入れた
